Time: 4 ms (25.98%) | Memory: 111.3 MB (73.84%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long ans = 0;
-        long long curcount  =0;
+        int curcount  =0;
         for(auto it : nums){
             if(it==0){
                 curcount++;
